services/catalog/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
motor>=3.3.1
nats-py>=2.3.1
pydantic>=2.0.0
PyJWT
